Move celery to services module
