+++ opwen_email_server/api/client_write.py
@@ -7,7 +7,7 @@ from opwen_email_server.services.auth import AzureAuth
 from opwen_email_server.services.storage import AzureTextStorage
 from opwen_email_server.utils.log import LogMixin
 
-from opwen_email_server.celery import celery
+from opwen_email_server.services import celery
 
 CLIENTS = AzureAuth(
     storage=AzureTextStorage(
